==> price_monitor/01_Price_today.ipynb <==
# Dependencies
from datetime import datetime
import datetime as dt
from pytz import timezone
import pandas as pd
import numpy as np

# Making time format
date_format = "%Y-%m-%d"
time_format = "%H:%M"
date_time_format = "%Y-%m-%d %H:%M"
year_format = "%y"
week_format = "Wk%y%W"

# Getting the current date & time
now_pst = datetime.now(timezone('US/Pacific'))
tomorrow_pst = now_pst + dt.timedelta(days=1)

# Getting ISO weeknumber and convert the format
week_data = tomorrow_pst.isocalendar()[1]
year_data = now_pst.strftime(year_format)

if week_data < 10 :
    week_data = "Wk" + year_data + "0" + str(week_data)
else: 
    week_data = "Wk" + year_data + str(week_data)
    
week_data

csv_path = "Resources/180830_price_master.csv"

# Import the books.csv file as a DataFrame
price_master_df = pd.read_csv(csv_path, encoding="utf-8")
price_master_df

price_master_today_df = price_master_df[["Product", week_data]]
price_master_today_df

price_master_today_df = price_master_today_df.rename(columns={week_data: f"Official Price {week_data}"})
price_master_today_df

price_master_today_df.to_csv("Output/price_today.csv", index=False)
