==> price_monitor/02_BB_Price_Monitor.ipynb <==
# Dependencies
from bs4 import BeautifulSoup
import requests
from datetime import datetime
from pytz import timezone
import pandas as pd
import numpy as np
from splinter import Browser
from splinter.exceptions import ElementDoesNotExist

# Make a reference to the books.csv file path
csv_path = "Resources/180826_Best_Buy_ProductId.csv"

# Import the books.csv file as a DataFrame
bby_productid_df = pd.read_csv(csv_path, encoding="utf-8")
bby_productid_df

for index, row in bby_productid_df.iterrows():
    url = row["producturl"]
    
    # Opening the page with chromedriver
    executable_path = {'executable_path': 'chromedriver.exe'}
    browser = Browser('chrome', **executable_path, headless=False)
    browser.visit(url)

    # Getting soup
    html = browser.html
    soup = BeautifulSoup(html, 'html.parser')
    
    # Getting price from soup
    try:
        price_info = soup.find('div', class_='activated-pricing').text
        price = float(price_info.split("(Activate Later)")[0].split("$")[-1].replace(' ', ''))
    except:
        try:
            price_info = soup.find('div', class_='priceView-hero-price priceView-customer-price').find_all('span')[0].text
            price = float(price_info.replace("$", ""))
        except:
            price = None
            
    bby_productid_df.at[index, "Best Buy"] = price

bby_productid_df

bby_result_df = bby_productid_df.drop(columns=["producturl"])
bby_result_df

bby_result_df.to_csv("Output/bby_result.csv", index=False)
