==> price_monitor/04_AZ_Price_Monitor.ipynb <==
# Dependencies
import requests
import bs4
from datetime import datetime
from pytz import timezone
import pandas as pd
import numpy as np
from urllib.request import Request, urlopen

from selenium import webdriver
from selenium.webdriver.chrome.options import Options
import time

# Make a reference to the books.csv file path
csv_path = "Resources/180818_Amazon_ProductId.csv"

# Import the books.csv file as a DataFrame
amazon_productid_df = pd.read_csv(csv_path, encoding="utf-8")
amazon_productid_df

amazon_productid_df["asin"] = amazon_productid_df["asin"].astype(object)

amazon_asin_list = amazon_productid_df["asin"]

master_df = None

for asin in amazon_asin_list:

    
    options = Options()

    options.set_headless(True)

    driver = webdriver.Chrome(executable_path='chromedriver.exe')

    driver.get(f"https://www.amazon.com/gp/offer-listing/{asin}/ref=dp_olp_all_mbc?_encoding=UTF8&condition=new")
    
    time.sleep(1)

    html = driver.page_source.encode('utf-8')

    soup = bs4.BeautifulSoup(html, "html.parser")
    
    offer_blocks = soup.select('.a-row.a-spacing-mini.olpOffer')
       
    # Making empty lists
    seller_list = []
    price_list = []
        
    # Making seller_list
    for offer in offer_blocks:
    
        offer_seller = offer.select('.a-spacing-none.olpSellerName')
        seller = offer_seller[0].getText().strip()
        if seller == "":
            seller = "Amazon.com"
        seller = seller + " in Amazon" 
    
        offer_price = offer.select('.a-size-large.a-color-price.olpOfferPrice.a-text-bold')
        if offer_price == []:
            price = ""
        else:
            price = offer_price[0].getText().strip()
            price = price.replace("$", "")
            price = price.replace(",", "")
            price = float(price)
        
        if seller in seller_list:
            continue
        
        else:
            seller_list.append(seller)
            price_list.append(price)
                
    product_price_df = pd.DataFrame({"Seller": seller_list,
                                         asin: price_list})

    product_price_df = product_price_df.set_index("Seller")

    master_df = pd.concat([master_df, product_price_df], axis=1, sort=False)

master_df = master_df.T
master_df = master_df.reset_index()
master_df = master_df.rename(columns={"index": "asin"})
master_df["asin"] = master_df["asin"].astype(object)
master_df = pd.merge(amazon_productid_df, master_df, on="asin", how="left")
master_df = master_df.rename(columns={"asin": "ASIN"})
master_df

amazon_result_df = master_df.drop(columns=["ASIN"])
amazon_result_df

amazon_result_df.to_csv("Output/amazon_result.csv", index=False)
