==> price_monitor/05_WM_Price_Monitor.ipynb <==
# Dependencies
import requests
from bs4 import BeautifulSoup
from datetime import datetime
from pytz import timezone
import pandas as pd
import numpy as np

# Make a reference to the books.csv file path
csv_path = "Resources/180818_Walmart_productId.csv"

# Import the books.csv file as a DataFrame
Walmart_productid_df = pd.read_csv(csv_path, encoding="utf-8")
Walmart_productid_df

Walmart_productid_df["productId"] = Walmart_productid_df["productId"].astype(object)

Walmart_product_id_list = Walmart_productid_df["productId"]

master_df = None

for product_id in Walmart_product_id_list:
    
    #Making soup
    response = requests.get(f"https://www.walmart.com/product/{product_id}/sellers")
    soup = BeautifulSoup(response.text, 'html.parser')
    
    #Making seller blocks
    results = soup.find_all('div', class_="product-seller-card")
    
    # Making empty lists
    seller_list = []
    price_list = []
    
    for result in results:
        try:
            seller = result.find_all('a', class_="seller-shipping-msg")[0].text
            seller = seller + " in Walmart"
            seller_list.append(seller)
        except:
            seller_list.append(None)
               
        try:
            price = result.find_all(class_="price-group")[0].text
            price = price.replace("$", "")
            price = float(price)
            price_list.append(price)
        except:
            price_list.append(None)
        
           
    product_price_df = pd.DataFrame({"Seller": seller_list,
                                     product_id: price_list})
    product_price_df = product_price_df.set_index("Seller")
    
    master_df = pd.concat([master_df, product_price_df], axis=1, sort=False)

master_df = master_df.T
master_df = master_df.reset_index()
master_df = master_df.rename(columns={"index": "productId"})
master_df["productId"] = master_df["productId"].astype(object)
master_df = pd.merge(Walmart_productid_df, master_df, on="productId", how="left")
master_df

walmart_result_df = master_df.drop(columns=["productId"])
walmart_result_df

walmart_result_df.to_csv("Output/walmart_result.csv", index=False)
