==> price_monitor/06_NE_Price_Monitor.ipynb <==
# Dependencies
import requests
import bs4
from datetime import datetime
from pytz import timezone
import pandas as pd
import numpy as np
from urllib.request import Request, urlopen
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
import time

# Make a reference to the books.csv file path
csv_path = "Resources/180826_Newegg_ProductId.csv"

# Import the books.csv file as a DataFrame
newegg_productid_df = pd.read_csv(csv_path, encoding="utf-8")
newegg_productid_df

newegg_productid_df["productid"] = newegg_productid_df["productid"].astype(object)

newegg_product_id_list = newegg_productid_df["productid"]

master_df = None

for productid in newegg_product_id_list:
    
    options = Options()

    options.set_headless(True)

    driver = webdriver.Chrome(executable_path='chromedriver.exe')

    driver.get(f"https://www.newegg.com/Product/Product.aspx?Item={productid}")
    
    time.sleep(1)

    html = driver.page_source.encode('utf-8')

    soup = bs4.BeautifulSoup(html, "html.parser")
    
    # Making empty lists
    seller_list = []
    price_list = []
    
    try:     
        seller = soup.select(".featured-seller")[0].getText()[21:]
        if seller.find("Sold and Shipped by") > 0:
            seller = seller[:seller.find("Sold and Shipped by")]
            seller = seller + " in Newegg.com"
            seller_list.append(seller)
        else:
            seller = seller + " in Newegg.com"
            seller_list.append(seller)
    except:
        seller = seller + " in Newegg.com"
        seller_list.append(seller)
   
        
    if soup.select(".price-current.hidden") != []:
        price = soup.select(".price-current.hidden")[0].getText()
        price = price.replace("$", "")
        price = float(price)
        price_list.append(price)
    
    elif soup.select("span.price-current") == []:
        price = ""
        price_list.append(price)    
    
    else:
        price = soup.select("span.price-current")[0].getText()
        price = price.replace("$", "")
        price = float(price)
        price_list.append(price)

    product_price_df = pd.DataFrame({"Seller": seller_list,
                                     productid: price_list})
    
    product_price_df = product_price_df.set_index("Seller")
    
    master_df = pd.concat([master_df, product_price_df], axis=1, sort=False)

master_df = master_df.T
master_df = master_df.reset_index()
master_df = master_df.rename(columns={"index": "productid"})
master_df["productid"] = master_df["productid"].astype(object)
master_df = pd.merge(newegg_productid_df, master_df, on="productid", how="left")
master_df

newegg_result_df = master_df.drop(columns=["productid"])
newegg_result_df

newegg_result_df.to_csv("Output/newegg_result.csv", index=False)
