==> price_monitor/07_Price_Monitor_Summary.ipynb <==
# Dependencies
import requests
import bs4
from datetime import datetime
from pytz import timezone
import pandas as pd
import numpy as np
import openpyxl
from openpyxl.styles.fonts import Font
from openpyxl.styles import PatternFill, Border, Side, Alignment, Protection, Font
from openpyxl import Workbook

fmt_1 = "%Y-%m-%d"
fmt_2 = "%H:%M"
fmt_F = "%H%M"

now_pst = datetime.now(timezone('US/Pacific'))

now_date_fmt = now_pst.strftime(fmt_1)
now_time_fmt = now_pst.strftime(fmt_2)
now_time_fmt_F = now_pst.strftime(fmt_F)
now_time_fmt_F

# Path
price_path = "Output/price_today.csv"
bby_path = "Output/bby_result.csv"
bh_path = "Output/bh_result.csv"
amazon_path = "Output/amazon_result.csv"
walmart_path = "Output/walmart_result.csv"
newegg_path = "Output/newegg_result.csv"

price_df = pd.read_csv(price_path, encoding="utf-8")
price_df = price_df.set_index("Product")
price_df

bby_df = pd.read_csv(bby_path, encoding="utf-8")
bby_df = bby_df.set_index("Product")
bby_df

bh_df = pd.read_csv(bh_path, encoding="utf-8")
bh_df = bh_df.set_index("Product")
bh_df

amazon_df = pd.read_csv(amazon_path, encoding="utf-8")
amazon_df = amazon_df.set_index("Product")
amazon_df

walmart_df = pd.read_csv(walmart_path, encoding="utf-8")
walmart_df = walmart_df.set_index("Product")
walmart_df

newegg_df = pd.read_csv(newegg_path, encoding="utf-8")
newegg_df = newegg_df.set_index("Product")
newegg_df

master_df = None

master_df = pd.concat([master_df, price_df], axis=1, sort=False)

master_df = pd.concat([master_df, bby_df], axis=1, sort=False)

master_df = pd.concat([master_df, bh_df], axis=1, sort=False)

master_df = pd.concat([master_df, walmart_df], axis=1, sort=False)

master_df = pd.concat([master_df, newegg_df], axis=1, sort=False)

master_df = pd.concat([master_df, amazon_df], axis=1, sort=False)

master_df = master_df.reset_index()
master_df = master_df.rename(columns={"index": "Product"})
master_df

master_df.to_csv(f"Output/{now_date_fmt}_{now_time_fmt_F}_Price_Monitor.csv", index=False)
master_df.to_excel(f"Output/{now_date_fmt}_{now_time_fmt_F}_Price_Monitor.xlsx", index=False)

wb =openpyxl.load_workbook(f'Output/{now_date_fmt}_{now_time_fmt_F}_Price_Monitor.xlsx')

sheet = wb['Sheet1']

sheet.column_dimensions['A'].width = 28

sheet.column_dimensions['B'].width = 10
sheet.column_dimensions['C'].width = 10
sheet.column_dimensions['D'].width = 10
sheet.column_dimensions['E'].width = 10
sheet.column_dimensions['F'].width = 10
sheet.column_dimensions['G'].width = 10
sheet.column_dimensions['H'].width = 10
sheet.column_dimensions['I'].width = 10
sheet.column_dimensions['J'].width = 10
sheet.column_dimensions['K'].width = 10
sheet.column_dimensions['L'].width = 10
sheet.column_dimensions['M'].width = 10
sheet.column_dimensions['N'].width = 10
sheet.column_dimensions['O'].width = 10
sheet.column_dimensions['P'].width = 10
sheet.column_dimensions['Q'].width = 10
sheet.column_dimensions['R'].width = 10
sheet.column_dimensions['S'].width = 10
sheet.column_dimensions['T'].width = 10
sheet.column_dimensions['U'].width = 10
sheet.column_dimensions['V'].width = 10
sheet.column_dimensions['W'].width = 10
sheet.column_dimensions['X'].width = 10
sheet.column_dimensions['Y'].width = 10
sheet.column_dimensions['Z'].width = 10
sheet.column_dimensions['AA'].width = 10
sheet.column_dimensions['AB'].width = 10
sheet.column_dimensions['AC'].width = 10
sheet.column_dimensions['AD'].width = 10
sheet.column_dimensions['AE'].width = 10
sheet.column_dimensions['AF'].width = 10
sheet.column_dimensions['AG'].width = 10
sheet.column_dimensions['AH'].width = 10
sheet.column_dimensions['AI'].width = 10
sheet.column_dimensions['AJ'].width = 10
sheet.column_dimensions['AK'].width = 10
sheet.column_dimensions['AL'].width = 10
sheet.column_dimensions['AM'].width = 10
sheet.column_dimensions['AN'].width = 10
sheet.column_dimensions['AO'].width = 10
sheet.column_dimensions['AP'].width = 10
sheet.column_dimensions['AQ'].width = 10
sheet.column_dimensions['AR'].width = 10
sheet.column_dimensions['AS'].width = 10
sheet.column_dimensions['AT'].width = 10
sheet.column_dimensions['AU'].width = 10
sheet.column_dimensions['AV'].width = 10
sheet.column_dimensions['AW'].width = 10
sheet.column_dimensions['AX'].width = 10
sheet.column_dimensions['AY'].width = 10
sheet.column_dimensions['AZ'].width = 10

sheet.freeze_panes = 'C2'

fill_1 = openpyxl.styles.PatternFill(patternType='solid', fgColor='c6e2ff', bgColor='c6e2ff')
fill_2 = openpyxl.styles.PatternFill(patternType='solid', fgColor='daedfe', bgColor='daedfe')

for i in range(1,53):
    sheet.cell(row=1, column=i).fill = fill_1
    sheet.cell(row=1, column=i).alignment = Alignment(horizontal='center', vertical='center', wrap_text=True)
    
    for x in range(2, 242):
        if x % 2 == 1:
            sheet.cell(row=x, column=i).fill = fill_2

for r in range(1, 242):
    for c in range(1, 53):
        sheet.cell(row=r, column=c).border = Border(outline=True, 
                                                    left=Side(style='dotted', color='FF000000'),
                                                    right=Side(style='dotted', color='FF000000'), 
                                                    top=Side(style='dotted', color='FF000000'), 
                                                    bottom=Side(style='dotted', color='FF000000'))
  
sheet.sheet_view.showGridLines = False

for r in range(2, 242):
    for i in range(2,53):
        if sheet.cell(row=r, column=i) != "":
            sheet.cell(row=r, column=i).number_format = '$  #,##0.00' 

for r in range(2, 242):
    for i in range(3,53):
        if sheet.cell(row=r, column=i).value == None:
            continue
        
        elif sheet.cell(row=r, column=2).value == None:
            continue
              
        elif sheet.cell(row=r, column=i).value < sheet.cell(row=r, column=2).value:
            sheet.cell(row=r, column=i).font = Font(color="FF0000", bold=True)            
            
for r in range(2, 242):
    sheet.cell(row=r, column=1).font = Font(bold=True)
    sheet.cell(row=r, column=2).font = Font(bold=True)

sheet.sheet_view.zoomScale = 85    
    
wb.save(f'Output/{now_date_fmt}_{now_time_fmt_F}_Price_Monitor.xlsx')
